backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
mysql-connector-python==8.1.0
sqlalchemy==2.0.21
flask-sqlalchemy==3.0.5
groq==0.4.1
marshmallow==3.20.1
